requirements.txt
pyzmq
pyaudio
numpy